382095919: (feat) add settings basic page
