+++ modules/ui/src/app/pages/general-settings/components/settings-dropdown/settings-dropdown.component.scss
