+++ modules/ui/src/app/pages/general-settings/general-settings.component.scss
@@ -1,0 +1,152 @@
+/**
+ * Copyright 2023 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+@use '@angular/material' as mat;
+@use 'colors';
+@use 'variables';
+
+:host {
+  display: flex;
+  flex-direction: column;
+  height: 100%;
+  flex: 1 0 auto;
+}
+
+.settings-drawer-header {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 12px 12px 16px 24px;
+
+  &-title {
+    margin: 0;
+    font-size: 22px;
+    font-style: normal;
+    font-weight: 400;
+    line-height: 28px;
+    color: colors.$dark-grey;
+  }
+
+  &-button {
+    min-width: 24px;
+    width: 24px;
+    height: 24px;
+    margin: 4px;
+    padding: 8px;
+    box-sizing: content-box;
+    line-height: normal !important;
+
+    .close-button-icon {
+      width: 24px;
+      height: 24px;
+      margin: 0;
+    }
+
+    ::ng-deep * {
+      line-height: inherit !important;
+    }
+  }
+}
+
+.setting-drawer-content {
+  padding: 0 16px 8px 16px;
+  overflow: hidden;
+  flex: 1;
+
+  form {
+    display: grid;
+    height: 100%;
+  }
+
+  .setting-drawer-content-form-empty {
+    grid-template-rows: repeat(2, auto) 1fr;
+  }
+}
+
+.setting-drawer-content-inputs {
+  overflow: auto;
+  margin: 0 -16px;
+  padding: 0 16px;
+}
+
+.error-message-container {
+  display: block;
+  margin-top: auto;
+  padding-bottom: 8px;
+}
+
+.error-message-container + .setting-drawer-footer {
+  margin-top: 0;
+}
+
+.message {
+  margin: 0;
+  padding: 6px 0 12px 0;
+  color: colors.$grey-800;
+  font-family: variables.$font-secondary;
+  font-size: 14px;
+  line-height: 20px;
+  letter-spacing: 0.2px;
+}
+
+.setting-drawer-footer {
+  padding: 0 8px;
+  margin-top: auto;
+  display: flex;
+  flex-shrink: 0;
+  justify-content: flex-end;
+
+  .close-button,
+  .save-button {
+    padding: 0 24px;
+    font-size: 14px;
+    font-weight: 500;
+    line-height: 20px;
+    letter-spacing: 0.25px;
+  }
+
+  .close-button {
+    margin-right: 10px;
+  }
+}
+
+.settings-disabled-overlay {
+  position: absolute;
+  width: 100%;
+  left: 0;
+  right: 0;
+  top: 75px;
+  bottom: 45px;
+  background-color: rgba(255, 255, 255, 0.7);
+  z-index: 2;
+}
+
+.disabled {
+  .message-link {
+    cursor: default;
+    pointer-events: none;
+
+    &:focus-visible {
+      outline: none;
+    }
+  }
+}
+
+.settings-drawer-header-button:not(.mat-mdc-button-disabled),
+.close-button:not(.mat-mdc-button-disabled),
+.save-button:not(.mat-mdc-button-disabled) {
+  cursor: pointer;
+  pointer-events: auto;
+}
