+++ modules/ui/src/app/pages/general-settings/only-different-values.validator.ts
