+++ modules/ui/src/app/pages/settings/settings.component.scss
@@ -13,140 +13,28 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-@use '@angular/material' as mat;
-@use 'colors';
-@use 'variables';
 
-:host {
-  display: flex;
-  flex-direction: column;
-  height: 100%;
-  flex: 1 0 auto;
+.toolbar {
+  height: auto;
+  padding: 22px 0px 18px 32px;
+}
+.title {
+  padding: 24px 0 16px;
 }
 
-.settings-drawer-header {
+.tab-item {
   display: flex;
-  justify-content: space-between;
+  padding: 0px 32px;
   align-items: center;
-  padding: 12px 12px 16px 24px;
-
-  &-title {
-    margin: 0;
-    font-size: 22px;
-    font-style: normal;
-    font-weight: 400;
-    line-height: 28px;
-    color: colors.$dark-grey;
+  gap: 32px;
+}
+.tab-group {
+  ::ng-deep .mat-mdc-tab-labels {
+    gap: 16px;
+    padding: 0 12px;
   }
 
-  &-button {
-    min-width: 24px;
-    width: 24px;
-    height: 24px;
-    margin: 4px;
-    padding: 8px;
-    box-sizing: content-box;
-    line-height: normal !important;
-
-    .close-button-icon {
-      width: 24px;
-      height: 24px;
-      margin: 0;
-    }
-
-    ::ng-deep * {
-      line-height: inherit !important;
-    }
+  ::ng-deep.mat-mdc-tab {
+    padding: 0px 8px;
   }
-}
-
-.setting-drawer-content {
-  padding: 0 16px 8px 16px;
-  overflow: hidden;
-  flex: 1;
-
-  form {
-    display: grid;
-    height: 100%;
-  }
-
-  .setting-drawer-content-form-empty {
-    grid-template-rows: repeat(2, auto) 1fr;
-  }
-}
-
-.setting-drawer-content-inputs {
-  overflow: auto;
-  margin: 0 -16px;
-  padding: 0 16px;
-}
-
-.error-message-container {
-  display: block;
-  margin-top: auto;
-  padding-bottom: 8px;
-}
-
-.error-message-container + .setting-drawer-footer {
-  margin-top: 0;
-}
-
-.message {
-  margin: 0;
-  padding: 6px 0 12px 0;
-  color: colors.$grey-800;
-  font-family: variables.$font-secondary;
-  font-size: 14px;
-  line-height: 20px;
-  letter-spacing: 0.2px;
-}
-
-.setting-drawer-footer {
-  padding: 0 8px;
-  margin-top: auto;
-  display: flex;
-  flex-shrink: 0;
-  justify-content: flex-end;
-
-  .close-button,
-  .save-button {
-    padding: 0 24px;
-    font-size: 14px;
-    font-weight: 500;
-    line-height: 20px;
-    letter-spacing: 0.25px;
-  }
-
-  .close-button {
-    margin-right: 10px;
-  }
-}
-
-.settings-disabled-overlay {
-  position: absolute;
-  width: 100%;
-  left: 0;
-  right: 0;
-  top: 75px;
-  bottom: 45px;
-  background-color: rgba(255, 255, 255, 0.7);
-  z-index: 2;
-}
-
-.disabled {
-  .message-link {
-    cursor: default;
-    pointer-events: none;
-
-    &:focus-visible {
-      outline: none;
-    }
-  }
-}
-
-.settings-drawer-header-button:not(.mat-mdc-button-disabled),
-.close-button:not(.mat-mdc-button-disabled),
-.save-button:not(.mat-mdc-button-disabled) {
-  cursor: pointer;
-  pointer-events: auto;
 }
