+++ modules/ui/src/app/pages/settings/settings.component.spec.ts
@@ -1,3 +1,6 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { SettingsComponent } from './settings.component';
 /**
  * Copyright 2023 Google LLC
  *
@@ -13,369 +16,23 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { NoopAnimationsModule } from '@angular/platform-browser/animations';
 
-import { SettingsComponent } from './settings.component';
-import { of } from 'rxjs';
-import { MatRadioModule } from '@angular/material/radio';
-import { ReactiveFormsModule } from '@angular/forms';
-import { MatButtonModule } from '@angular/material/button';
-import { MatIcon, MatIconModule } from '@angular/material/icon';
-import { MatIconTestingModule } from '@angular/material/icon/testing';
-import { Component, Input } from '@angular/core';
-import { LiveAnnouncer } from '@angular/cdk/a11y';
-import SpyObj = jasmine.SpyObj;
-import { MatInputModule } from '@angular/material/input';
-import { MatSelectModule } from '@angular/material/select';
-import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { provideMockStore } from '@ngrx/store/testing';
-import { LoaderService } from '../../services/loader.service';
-import { SettingsStore } from './settings.store';
-import {
-  MOCK_INTERFACES,
-  MOCK_SYSTEM_CONFIG_WITH_DATA,
-} from '../../mocks/settings.mock';
-import { SettingsDropdownComponent } from './components/settings-dropdown/settings-dropdown.component';
-import { CalloutComponent } from '../../components/callout/callout.component';
-import { SpinnerComponent } from '../../components/spinner/spinner.component';
-
-describe('GeneralSettingsComponent', () => {
+describe('SettingsComponent', () => {
   let component: SettingsComponent;
   let fixture: ComponentFixture<SettingsComponent>;
-  let mockLiveAnnouncer: SpyObj<LiveAnnouncer>;
-  let compiled: HTMLElement;
-  let mockLoaderService: SpyObj<LoaderService>;
-  let mockSettingsStore: SpyObj<SettingsStore>;
 
   beforeEach(async () => {
-    mockLiveAnnouncer = jasmine.createSpyObj(['announce']);
-    mockLoaderService = jasmine.createSpyObj('LoaderService', ['setLoading']);
-    mockSettingsStore = jasmine.createSpyObj('SettingsStore', [
-      'getInterfaces',
-      'updateSystemConfig',
-      'setIsSubmitting',
-      'setDefaultFormValues',
-      'getSystemConfig',
-      'viewModel$',
-    ]);
-
     await TestBed.configureTestingModule({
-      providers: [
-        { provide: LiveAnnouncer, useValue: mockLiveAnnouncer },
-        { provide: LoaderService, useValue: mockLoaderService },
-        { provide: SettingsStore, useValue: mockSettingsStore },
-        provideMockStore(),
-      ],
-      imports: [
-        SettingsComponent,
-        BrowserAnimationsModule,
-        MatButtonModule,
-        MatIconModule,
-        MatRadioModule,
-        ReactiveFormsModule,
-        MatIconTestingModule,
-        MatIcon,
-        MatInputModule,
-        MatSelectModule,
-        SettingsDropdownComponent,
-        FakeSpinnerComponent,
-        FakeCalloutComponent,
-      ],
-    })
-      .overrideComponent(SettingsComponent, {
-        remove: {
-          imports: [CalloutComponent, SpinnerComponent],
-        },
-        add: {
-          imports: [FakeSpinnerComponent, FakeCalloutComponent],
-        },
-      })
-      .compileComponents();
-
-    TestBed.overrideProvider(SettingsStore, { useValue: mockSettingsStore });
+      imports: [SettingsComponent, NoopAnimationsModule],
+    }).compileComponents();
 
     fixture = TestBed.createComponent(SettingsComponent);
-
     component = fixture.componentInstance;
-    component.viewModel$ = of({
-      systemConfig: { network: {} },
-      hasConnectionSettings: false,
-      isSubmitting: false,
-      isLessThanOneInterface: false,
-      interfaces: {},
-      deviceOptions: {},
-      internetOptions: {},
-      logLevelOptions: {},
-      monitoringPeriodOptions: {},
-    });
     fixture.detectChanges();
-    compiled = fixture.nativeElement as HTMLElement;
-
-    component.ngOnInit();
   });
 
   it('should create', () => {
     expect(component).toBeTruthy();
   });
-
-  it('#reloadSetting should call setLoading in loaderService', () => {
-    component.reloadSetting();
-
-    expect(mockLoaderService.setLoading).toHaveBeenCalledWith(true);
-  });
-
-  describe('#settingsDisable', () => {
-    it('should disable setting form when get settingDisable as true ', () => {
-      spyOn(component.settingForm, 'disable');
-
-      component.settingsDisable = true;
-
-      expect(component.settingForm.disable).toHaveBeenCalled();
-    });
-
-    it('should enable setting form when get settingDisable as false ', () => {
-      spyOn(component.settingForm, 'enable');
-
-      component.settingsDisable = false;
-
-      expect(component.settingForm.enable).toHaveBeenCalled();
-    });
-
-    it('should disable "Save" button when get settingDisable as true', () => {
-      component.settingsDisable = true;
-
-      const saveBtn = compiled.querySelector(
-        '.save-button'
-      ) as HTMLButtonElement;
-
-      expect(saveBtn.disabled).toBeTrue();
-    });
-
-    it('should disable "Refresh" link when settingDisable', () => {
-      component.settingsDisable = true;
-
-      const refreshLink = compiled.querySelector(
-        '.message-link'
-      ) as HTMLAnchorElement;
-
-      refreshLink.click();
-
-      expect(refreshLink.hasAttribute('aria-disabled')).toBeTrue();
-      expect(mockLoaderService.setLoading).not.toHaveBeenCalled();
-    });
-  });
-
-  describe('#closeSetting', () => {
-    beforeEach(() => {
-      component.ngOnInit();
-    });
-
-    it('should emit closeSettingEvent', () => {
-      spyOn(component.closeSettingEvent, 'emit');
-
-      component.closeSetting('Message');
-
-      expect(component.closeSettingEvent.emit).toHaveBeenCalled();
-    });
-
-    it('should call liveAnnouncer with provided message', () => {
-      const mockMessage = 'mock event';
-
-      component.closeSetting(mockMessage);
-
-      expect(mockLiveAnnouncer.announce).toHaveBeenCalledWith(
-        `The ${mockMessage} finished. The system settings panel is closed.`
-      );
-    });
-
-    it('should call reset settingForm', () => {
-      spyOn(component.settingForm, 'reset');
-
-      component.closeSetting('Message');
-
-      expect(component.settingForm.reset).toHaveBeenCalled();
-    });
-
-    it('should call setDefaultFormValues', () => {
-      component.closeSetting('Message');
-
-      expect(mockSettingsStore.setDefaultFormValues).toHaveBeenCalled();
-    });
-  });
-
-  describe('#saveSetting', () => {
-    beforeEach(() => {
-      component.ngOnInit();
-    });
-
-    it('should have form error if form has the same value', () => {
-      const mockSameValue = 'sameValue';
-      component.deviceControl.setValue(mockSameValue);
-      component.internetControl.setValue(mockSameValue);
-
-      component.saveSetting();
-
-      expect(component.settingForm.invalid).toBeTrue();
-      expect(component.isFormError).toBeTrue();
-      expect(mockSettingsStore.setIsSubmitting).toHaveBeenCalledWith(true);
-    });
-
-    it('should call createSystemConfig when setting form valid', () => {
-      const expectedResult = {
-        network: {
-          device_intf: 'mockDeviceKey',
-          internet_intf: '',
-        },
-        log_level: 'INFO',
-        monitor_period: 600,
-      };
-
-      component.deviceControl.setValue({
-        key: 'mockDeviceKey',
-        value: 'mockDeviceValue',
-      });
-
-      component.internetControl.setValue({
-        key: '',
-        value: 'defaultValue',
-      });
-
-      component.logLevel.setValue({
-        key: 'INFO',
-        value: '',
-      });
-
-      component.monitorPeriod.setValue({
-        key: '600',
-        value: '',
-      });
-
-      component.saveSetting();
-
-      const args = mockSettingsStore.updateSystemConfig.calls.argsFor(0);
-      // @ts-expect-error config is in object
-      expect(args[0].config).toEqual(expectedResult);
-      expect(component.settingForm.invalid).toBeFalse();
-      expect(mockSettingsStore.updateSystemConfig).toHaveBeenCalled();
-    });
-  });
-
-  describe('with no interfaces data', () => {
-    beforeEach(() => {
-      component.viewModel$ = of({
-        systemConfig: { network: {} },
-        hasConnectionSettings: false,
-        isSubmitting: false,
-        isLessThanOneInterface: false,
-        interfaces: {},
-        deviceOptions: {},
-        internetOptions: {},
-        logLevelOptions: {},
-        monitoringPeriodOptions: {},
-      });
-      fixture.detectChanges();
-    });
-
-    it('should have callout component', () => {
-      const callout = compiled.querySelector('app-callout');
-
-      expect(callout).toBeTruthy();
-    });
-
-    it('should have disabled "Save" button', () => {
-      const saveBtn = compiled.querySelector(
-        '.save-button'
-      ) as HTMLButtonElement;
-
-      expect(saveBtn.disabled).toBeTrue();
-    });
-  });
-
-  describe('with interfaces length less than one', () => {
-    beforeEach(() => {
-      component.viewModel$ = of({
-        systemConfig: { network: {} },
-        hasConnectionSettings: false,
-        isSubmitting: false,
-        isLessThanOneInterface: true,
-        interfaces: {},
-        deviceOptions: {},
-        internetOptions: {},
-        logLevelOptions: {},
-        monitoringPeriodOptions: {},
-      });
-      fixture.detectChanges();
-    });
-
-    it('should have disabled "Save" button', () => {
-      component.deviceControl.setValue(
-        MOCK_SYSTEM_CONFIG_WITH_DATA?.network?.device_intf
-      );
-      component.internetControl.setValue(
-        MOCK_SYSTEM_CONFIG_WITH_DATA?.network?.internet_intf
-      );
-      fixture.detectChanges();
-
-      const saveBtn = compiled.querySelector(
-        '.save-button'
-      ) as HTMLButtonElement;
-
-      expect(saveBtn.disabled).toBeTrue();
-    });
-  });
-
-  describe('with interfaces length more then one', () => {
-    beforeEach(() => {
-      component.viewModel$ = of({
-        systemConfig: { network: {} },
-        hasConnectionSettings: false,
-        isSubmitting: false,
-        isLessThanOneInterface: false,
-        interfaces: MOCK_INTERFACES,
-        deviceOptions: MOCK_INTERFACES,
-        internetOptions: MOCK_INTERFACES,
-        logLevelOptions: {},
-        monitoringPeriodOptions: {},
-      });
-      fixture.detectChanges();
-    });
-
-    it('should not have callout component', () => {
-      const callout = compiled.querySelector('app-callout');
-
-      expect(callout).toBeFalsy();
-    });
-
-    it('should not have disabled "Save" button', () => {
-      component.deviceControl.setValue({
-        key: MOCK_SYSTEM_CONFIG_WITH_DATA?.network?.device_intf,
-        value: 'value',
-      });
-      component.internetControl.setValue({
-        key: MOCK_SYSTEM_CONFIG_WITH_DATA?.network?.internet_intf,
-        value: 'value',
-      });
-      fixture.detectChanges();
-
-      const saveBtn = compiled.querySelector(
-        '.save-button'
-      ) as HTMLButtonElement;
-
-      expect(saveBtn.disabled).toBeFalse();
-    });
-  });
 });
-
-@Component({
-  selector: 'app-spinner',
-  template: '<div></div>',
-})
-class FakeSpinnerComponent {}
-
-@Component({
-  selector: 'app-callout',
-  template: '<div></div>',
-})
-class FakeCalloutComponent {
-  @Input() type = '';
-}
